moodappupdated070326-main/mood-classification-main/moodapp-newfile-main/moodapp-newfile-main/api/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
joblib==1.5.3
python-dotenv==1.2.1
requests==2.32.5
google-generativeai==0.8.5
numpy==1.26.4
opencv-python==4.8.1.78
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.3.3
SQLAlchemy==2.0.36
PyJWT==2.9.0
bcrypt==4.2.0
cryptography==43.0.1
pyotp==2.9.0
